backend/services/python-service/requirements.txt
moorcheh-sdk
fastapi
uvicorn
python-dotenv
